intro: remove two non-sequitur sentences
The point of this paragraph is to briefly state the relevant history, and to describe the current state of astropy.

A general statement about aims would belong further up (I'm not sure where, so I did not put it anywhere). The object of "these tools" in original wording was not clear.